requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
redis==5.0.8
boto3==1.35.14
python-dotenv==1.0.1
pydantic==2.9.2